// src/app/shared/models/Food.ts
export class Foods{
    id: number = 0;
    name: string = '';
    price: number = 0;
    cookTime: string = '';
    favorite: boolean = false;
    origins: string[] = [];
    stars: number = 0;
    imageUrl: string = '';
    tags: string[] = [];
}

export class Products{
    body: string = '';
    email: string = '';
    id: number = 0;
    name: string = '';
    postId: number = 0;
  }